Refactored code to align with the style-guides
