+++ frontend/src/common/suggestitheader/elements/about/index.js
@@ -1,0 +1,3 @@
+import About from "./about.element";
+
+export default About;
