+++ frontend/src/common/suggestitheader/index.js
@@ -1,0 +1,3 @@
+import SuggestITHeaderView from "./suggestitheader.view";
+
+export default SuggestITHeaderView;
