+++ frontend/src/redux/SuggestionStore.js
@@ -1,22 +1,22 @@
 import { createStore } from "redux";
+import {
+  ADD_SUGGESTION, 
+  SET_SUGGESTIONS
+} from "./suggestionstore.actions";
 
-const reducer = function(state, action) {
+const reducer = (state, action) => {
     if (action.suggestion != null)
         switch (action.type) {
-            case "set":
-                return [...action.suggestion].sort(function(b, a) {
-                    return (
+            case SET_SUGGESTIONS:
+                return [...action.suggestion].sort((b, a) => 
                         new Date(a.timestamp).getTime() -
                         new Date(b.timestamp).getTime()
                     );
-                });
-            case "add":
-                return state.concat(action.suggestion).sort(function(b, a) {
-                    return (
+            case ADD_SUGGESTION:
+                return state.concat(action.suggestion).sort((b, a) =>
                         new Date(a.timestamp).getTime() -
                         new Date(b.timestamp).getTime()
-                    );
-                });
+                );
             default:
                 return state;
         }
