+++ frontend/src/redux/suggestionstore.actions.js
@@ -1,0 +1,2 @@
+export const ADD_SUGGESTION = "add_suggestion";
+export const SET_SUGGESTIONS = "set_suggestions";
