+++ frontend/src/use-cases/common/SuggestionItem/index.js
@@ -1,0 +1,3 @@
+import SuggestionItem from "./suggestionitem.view";
+
+export default SuggestionItem;
