+++ frontend/src/use-cases/common/SuggestionItem/suggestionitem.style.css
