+++ frontend/src/use-cases/home/views/SuggestionBoard/SuggestionBoard.css
