+++ frontend/src/use-cases/notfound/index.js
@@ -1,0 +1,3 @@
+import Notfound from "./notfound.element";
+
+export default Notfound;
