+++ frontend/src/use-cases/suggestion/index.js
@@ -1,0 +1,3 @@
+import Suggestion from "./suggestion.screen";
+
+export default Suggestion;
