--- frontend/src/use-cases/test/Test.jsx
@@ -1,10 +1,0 @@
-import React from "react";
-
-export const Test = () => (
-    <div>
-        <h1>Hi!</h1>
-        <p> This code exists in frontend/src/use-cases/test/Test.jsx</p>
-    </div>
-);
-
-export default Test;
